cleanup

+++ atlas-node-api/src/main/java/de/atlasmc/map/MapIcon.java
@@ -5,7 +5,7 @@ public class MapIcon implements Cloneable {
 	private IconType type;
 	private int x;
 	private int z;
-	private int direction;
+	private float direction;
 	private String name;
 	
 	public MapIcon(IconType type) {
@@ -16,7 +16,7 @@ public class MapIcon implements Cloneable {
 		this(type, x, z, 0);
 	}
 	
-	public MapIcon(IconType type, int x, int z, int direction) {
+	public MapIcon(IconType type, int x, int z, float direction) {
 		this.type = type;
 		this.x = x;
 		this.z = z;
@@ -69,7 +69,7 @@ public class MapIcon implements Cloneable {
 		return z;
 	}
 
-	public int getDirection() {
+	public float getDirection() {
 		return direction;
 	}
 
@@ -89,7 +89,7 @@ public class MapIcon implements Cloneable {
 		this.z = z;
 	}
 
-	public void setDirection(int direction) {
+	public void setDirection(float direction) {
 		this.direction = direction;
 	}
